test: adding Proponent income validtor tests. And fixing a bug with the bussines code. That's why I love tests :D

+++ ProposalValidator.Domain/Models/Proponent.cs
@@ -10,7 +10,7 @@ namespace ProposalValidator.Domain.Models
 
         public int Age { get; set; }
 
-        public int MonthlyIncome { get; set; }
+        public decimal MonthlyIncome { get; set; }
 
         public bool IsMain { get; set; }
     }
